requirements.txt
fastapi[standard]
uvicorn[standard]
sqlalchemy
passlib
python-jose
psycopg2
bcrypt
test
stest
